more stuff

+++ src/app/shared/components/local-admin-column-config-button/local-admin-column-config-button.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { NotificationService } from '../../services/notification.service';
 import { ITSystemUsageActions } from 'src/app/store/it-system-usage/actions';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../dialogs/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-local-admin-column-config-button',
@@ -10,16 +12,40 @@ import { ITSystemUsageActions } from 'src/app/store/it-system-usage/actions';
 })
 export class LocalAdminColumnConfigButtonComponent {
 
-  constructor(private store: Store, private notificationService: NotificationService) {}
+  private readonly organizationName = "HENT KOMMUNENAVN HER";
+
+  constructor(private store: Store, private notificationService: NotificationService, private dialog: MatDialog) {}
 
   public onSave(): void {
-    this.store.dispatch(ITSystemUsageActions.saveOrganizationalITSystemUsageColumnConfiguration());
-    this.notificationService.showDefault($localize`Kolonneopsætningen er gemt for organisationen`);
+    this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        title: $localize`Gem kolonneopsætning`,
+        bodyText: $localize`Er du sikker på at du vil gemme nuværende kolonneopsætning af felter som standard til ` + this.organizationName,
+        confirmText: $localize`Gem`,
+        declineText: $localize`Annuller`,
+      },
+    }).afterClosed().subscribe((pressedConfirm) => {
+      if (pressedConfirm) {
+        this.store.dispatch(ITSystemUsageActions.saveOrganizationalITSystemUsageColumnConfiguration());
+        this.notificationService.showDefault($localize`Kolonneopsætningen er gemt for organisationen`);
+      }
+    });
   }
 
   public onDelete(): void {
-    this.store.dispatch(ITSystemUsageActions.deleteOrganizationalITSystemUsageColumnConfiguration());
-    this.notificationService.showDefault($localize`Organisationens kolonneopsætningen er slettet og overblikket er nulstillet`);
+    this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        title: $localize`Slet kolonneopsætning`,
+        bodyText: $localize`Er du sikker på at du vil slette standard kolonneopsætning af felter til ` + this.organizationName,
+        confirmText: $localize`Slet`,
+        declineText: $localize`Annuller`,
+      },
+    }).afterClosed().subscribe((pressedConfirm) => {
+      if (pressedConfirm) {
+        this.store.dispatch(ITSystemUsageActions.deleteOrganizationalITSystemUsageColumnConfiguration());
+        this.notificationService.showDefault($localize`Organisationens kolonneopsætningen er slettet og overblikket er nulstillet`);
+      }
+    });
   }
 
 }
